Last version

+++ src/main/java/charts/myCharts/OpAvgFuture_E2_IndexCounter_Index_Chart.java
@@ -32,7 +32,7 @@ public class OpAvgFuture_E2_IndexCounter_Index_Chart extends MyChartCreator {
         props.setProp(ChartPropsEnum.IS_LIVE, false);
         props.setProp(ChartPropsEnum.SLEEP, 1000);
         props.setProp(ChartPropsEnum.CHART_MAX_HEIGHT_IN_DOTS, (double) INFINITE);
-        props.setProp(ChartPropsEnum.SECONDS_ON_MESS, (double) INFINITE);
+        props.setProp(ChartPropsEnum.SECONDS_ON_MESS, INFINITE);
         props.setProp(ChartPropsEnum.INCLUDE_DOMAIN_AXIS, true);
 
         Exp e1 = client.getExps().getExp(ExpStrings.e1);
